requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
awsebcli==3.20.7
base58==2.1.1
bcrypt==4.0.1
billiard==4.1.0
bitarray==2.8.0
blessed==1.20.0
boto3==1.20.54
botocore==1.23.54
cached-property==1.5.2
celery==5.3.1
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
cryptography==41.0.2
cytoolz==0.12.2
Django==4.2.3
django-cors-headers==4.2.0
django-ebhealthcheck==2.0.2
django-health-check==3.17.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.6
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
inflection==0.5.1
ipfshttpclient==0.7.0a1
jmespath==0.10.0
jsonschema==3.2.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
packaging==23.1
paramiko==3.2.0
parsimonious==0.8.1
pathspec==0.10.1
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.18.0
pycurl==7.45.2
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pyunormalize==15.0.0
PyYAML==5.4.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rlp==2.0.1
rpds-py==0.9.2
s3transfer==0.5.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
toolz==0.12.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
varint==1.0.2
vine==5.0.0
wcwidth==0.1.9
web3==5.17.0
websocket-client==0.59.0
websockets==8.1
yarl==1.9.2
